requirements.txt
numpy==2.1.1
pandas==2.2.2
Pillow==10.4.0
Requests==2.32.3
tqdm==4.66.5
transformers==4.44.2
